pyservices/store/requirements.txt
grpcio==1.13.0
grpcio-tools==1.13.0
protobuf==3.6.0
googleapis-common-protos==1.5.3
six==1.11.0
jsonpatch==1.21
eventsourcing==6.0.0
psycopg2-binary==2.7.4
pytest==3.2.3
retrying==1.3.3
SQLAlchemy==1.2.7
SQLAlchemy-Utils==0.32.21
ipython==7.0.1